Fix errant opening of SQLite database in situ prior to generating a copy and fix iOS 26 being detected as iOS 18.
This commit fixes a logic error wherein the SQLite database was being opened to check if it was valid prior to copying the original to a temporary location which led to the error reported in Bug #126.
This commit also fixes iOS 26 being identified as iOS 18 and updates the RSpec data README accordingly.

Fixes #126.